Questioning initialization recommendation
Currently in the documentation, the recommendation for express
applications is to set the Bookshelf instance onto your application.  I
find this to be not only an odd choice, but also a limiting one.

In Node, if you want something memoized easily, the best way to achieve
that is through a module.  They are cached after the first invocation
and provide the exact solution described.

My problem with the current recommendation is that you create a
completely unnecesary dependency on the express app from your model
definition.  This is limiting in so many ways, especially with regards
to testing.

I don't think it hurts to attach the bookshelf instance to your app, and
that may make sense for a lot operations, but for Model definitions I
don't see why you'd ever want this.

What do you think?