Do not create pivot model on a `through` relation's `create()`.
Although this might be a desirable feature, it does not seem to work as intended. Currently when using `create` on a `hasMany(X).through(Y)` relationship, this creates an incorrect joining model (it puts x_id on Y rather than y_id on X). Additionally this feature is undocumented and has no test coverage, so it seems appropriate remove it (at least for now).